feat: unify ExhStatusCard and ExhInfoCard
