+++ src/components/GuestScan.tsx
@@ -17,12 +17,11 @@ import CardList from "components/CardList";
 import DirectInputFab from "components/DirectInputFab";
 import DirectInputModal from "components/DirectInputModal";
 import ErrorAlert from "components/ErrorAlert";
-import ExhInfoCard from "components/ExhInfoCard";
 import QRScanner from "components/QRScanner";
 import ResultChip, { ResultChipRefs } from "components/ResultChip";
+import { ExhStatusCard } from "components/StayStatusCard";
 import useErrorHandler from "hooks/useErrorHandler";
 import { StatusColor } from "types/statusColor";
-import { ExhStatusCard } from "components/StayStatusCard";
 
 const useStyles = makeStyles((theme) =>
   createStyles({
@@ -194,7 +193,6 @@ const GuestScan: React.VFC<Props> = ({ handleScan, page }) => {
               </CardContent>
             </Card>
             {isExh && <ExhStatusCard />}
-            {isExh && <ExhInfoCard />}
           </CardList>
         </Grid>
       </Grid>
